pyproject.toml
[project]
name = "cartograph"
version = "0.1.0"
description = "Cartograph — autonomous deep research agent, fully local, zero API keys"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langchain-community>=0.3.0",
    "ddgs>=1.0.0",
    "faiss-cpu>=1.9.0",
    "sentence-transformers>=3.4.0",
    "gradio>=5.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
